pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pr-agent"
version = "1.0.0"
description = "PR-Agent: Multi-agent system for GitHub Pull Request reviews using CrewAI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai>=0.36.0",
    "crewai-tools>=0.4.0",
    "PyGithub>=1.59.1",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "packaging>=21.0",
    "semver>=3.0.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[project.scripts]
pr-agent = "pr_agent.main:main"
